requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
torch>=2.0.0
transformers>=4.30.0
colpali-engine>=0.3.9
faiss-cpu>=1.7.3
langchain>=0.0.300
langchain-community>=0.2.0
bitarray>=2.7.0
scikit-learn>=1.2.0
beir>=1.0.0
vidore-benchmark